pyproject.toml
[tool.poetry.dependencies]
python = "^3.13"
roboticstoolbox-python = "*"
numpy = "==1.26.4"
tabulate = "*"